"Refactor(component tree): Change the structure of the component tree, plus folder and file names
Until now, I didn't have a clear folder structure and naming conventions for my folders and files.
I have therefore created a new structure and naming convention as follows:

- Renamed the 'components' folder to 'features'

- Grouped files by feature. UI features get their own folder

- If a feature is made up of more than one component, place these additional components in their
own folder in a subfolder named 'components'.
Example: ProductPage/components/Description/Description.jsx,
ProductPage/components/Carousel/Carousel.jsx

- Use PascalCase for all component files and folders. Use the .jsx file extension for component files

- Use hyphen-case for all non-component files, and for folders that only contain other folders
or non-component files

Made the following changes to folders and files:

- Changed all folder and file names to comply with the above naming convention

- Created a new folder, 'ui', and moved the folders 'Header', 'Modal', and 'Dashboard' there

- Renamed the files DashboardItem.js/.css to Chart.jsx/.css and moved them to their own folder:
src/features/Chart.

- Moved the 'highcharts' folder from src/features/highcharts to src/features/Chart/highcharts

- Renamed the folder SearchFunction to Search, and the files SearchFunction.js/.css to Search.jsx/.css

- Renamed the folder SearchResult to Result, and SearchResult.jsx/.css to Results.jsx/.css

- Moved the folder alpha-vantage from src/features/alpha-vantage to src/features/Search/api/alpha-vantage"
