+++ src/features/Search/api/alpha-vantage/alpha-vantage-key.js
@@ -1,0 +1,3 @@
+const apiKey = "PBKENT6HA5JNEIDY";
+
+export default apiKey;
