+++ src/features/Search/api/alpha-vantage/currency-lists/digital-currency-list.js
@@ -1,4 +1,4 @@
-import { physicalCurrencies } from "./PhysicalCurrencyList";
+import { physicalCurrencies } from "./physical-currency-list";
 
 const digitalCurrencies = [
   { name: "FirstBlood", symbol: "1ST" },
