+++ src/features/Search/components/Result/Result.css
